rename: add jwt package and move files
