--- src/main/java/com/example/backend/global/security/JwtType.java
@@ -1,5 +1,0 @@
-package com.example.backend.global.security;
-
-public enum JwtType {
-    ACCESS, REFRESH
-}
