+++ src/main/java/com/example/backend/global/security/jwt/JwtExceptionEntryPoint.java
@@ -1,16 +1,13 @@
-package com.example.backend.global.security;
+package com.example.backend.global.security.jwt;
 
 import com.example.backend.dto.ErrorDTO;
 import com.example.backend.global.exception.UnAuthorizedExceptionType;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import lombok.RequiredArgsConstructor;
-import org.springframework.http.HttpStatus;
 import org.springframework.security.core.AuthenticationException;
-import org.springframework.security.core.userdetails.UserDetailsService;
 import org.springframework.security.web.AuthenticationEntryPoint;
 import org.springframework.stereotype.Component;
 
-import javax.servlet.ServletException;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
