+++ src/main/java/com/example/backend/global/security/jwt/JwtType.java
@@ -1,0 +1,5 @@
+package com.example.backend.global.security.jwt;
+
+public enum JwtType {
+    ACCESS, REFRESH
+}
